fix bug5: because of a source file lookup error macro calls weren't correctly evaluated in the syntax highlighting. bug24 needs more work...

+++ Synopsis/Parsers/Cxx/Parser.py
@@ -66,9 +66,6 @@ class Parser(Processor):
                               self.syntax_prefix,
                               self.xref_prefix)
 
-         if self.preprocess:
-
-            os.remove(ii_file)
-
+         if self.preprocess: os.remove(ii_file)
 
       return self.output_and_return_ast()
